Fix matching for multiple terms/clauses
The way I was matching before would cause a single term to match as a single
hash and multiple terms to match as an array of hashes.

This led to bugs in the transformers because they didn't match a sequences when
there was a single term.

I changed the parsers so they always return an array for an element that can be
repeated. I also split up the tests into separate files and added more tests.