Integration tests for Elasticsearch query generation
Add a few tests that exercise the full query parse-generation-execution cycle to
validate the generated queries. Includes some helpers for inserting data into
Elasticsearch in order to have something to search over. The index is re-created
and the data re-indexed at the beginning of the test suite run.

To run these tests, Elasticsearch must be running.