Update code to have the same class names
I made the class names different to separate the code in to self-contained
examples. However, this made the tutorial harder to follow, because there's no
reason why the class names need to change. It also didn't fully work, because
requiring one file after another redefines the classes that did have the same
name (for example, common code like Operator or TermClause).

This change puts the example code in separate modules so they are truly
self-contained. Meanwhile, I can update my examples in the text to elide the
difference so it seems like we are iterating on a single example.